+++ ingq/db/redis_cache.py
@@ -1,7 +1,7 @@
 from typing import Optional
 import redis.asyncio as redis
 from redis.asyncio.client import Redis
-from core.redis_config import RedisConfig
+from config.redis_config import RedisConfig
 
 
 class RedisCache:
